+++ internal/permissions/dummy_permission_store.go
